Rename angler.py to __init__.py, add __main__.py
