+++ angler/__main__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python3
+from . import main
+
+main()
